reorg ip blocks
